requirements.txt
google-genai>=1.0.0
python-dotenv>=1.0.1
langchain>=1.2.15
fastapi>=0.115.0
uvicorn>=0.34.0
